+++ demo/layouts/tags.vue
@@ -1,3 +1,0 @@
-<template>
-  <TagsThemeFlaribbit />
-</template>