+++ theme/components/common/SakuraButton.vue
@@ -7,26 +7,29 @@ withDefaults(defineProps<{
 </script>
 
 <template>
-  <button :class="type">
+  <button class="sakura-button" :class="`sakura-button-${type}`">
     <slot />
   </button>
 </template>
 
 <style lang="scss">
-.primary {
+.sakura-button {
   border-width: 1px;
-  background-color: transparent;
-  color: var(--st-c-text);
-  border-radius: 20px;
   padding: 2px 20px;
+  border-radius: 20px;
+  color: var(--st-c-text);
   transition:
     color 0.3s ease-in-out,
     border-color 0.2s ease-in-out;
 
-  &:hover,
-  &.clicked {
-    color: var(--st-c-secondary);
-    border-color: var(--st-c-secondary);
+  &-primary {
+    background-color: transparent;
+
+    &:hover,
+    &.clicked {
+      color: var(--st-c-secondary);
+      border-color: var(--st-c-secondary);
+    }
   }
 }
 </style>