+++ theme/composables/tags.ts
@@ -1,23 +1,52 @@
 import { TinyColor } from '@ctrl/tinycolor'
 import { useTags } from 'valaxy'
+import { useThemeConfig } from './config'
 
 /**
  * get utils about tags
  */
 export function useSakuraTags(options: {
-  /**
-   * Primary Color
-   */
-  primary: string
+  rainbow: string[]
 } = {
-  primary: '#0078E7',
+  rainbow: [
+    '#ff4e6a',
+    '#ff761e',
+    '#ffb900',
+    '#33d57a',
+    '#00dbff',
+    '#1a98ff',
+    '#9090ff',
+  ],
 }) {
+  const themeConfig = useThemeConfig()
   const tags = useTags()
 
-  const gray = new TinyColor('#999999')
-  const primaryColor = new TinyColor(options.primary)
+  const getTagStyleByIndex = (index: number) => {
+    const rainbow = themeConfig.value.tags?.rainbow === true
+      ? options.rainbow
+      : Array.isArray(themeConfig.value.tags?.rainbow)
+        ? themeConfig.value.tags.rainbow
+        : false
 
-  const getTagStyle = (count: number) => {
+    if (!rainbow)
+      return
+
+    const colorIndex = index % rainbow.length
+    const baseColor = new TinyColor(rainbow[colorIndex])
+
+    const textColor = baseColor.toString()
+    const backgroundColor = baseColor.setAlpha(0.15).toString()
+
+    return {
+      '--sakura-tag-color': textColor,
+      '--sakura-tag-bg-color': backgroundColor,
+    }
+  }
+
+  const getTagStyleByCount = (count: number) => {
+    const gray = new TinyColor('#999999')
+    const primaryColor = new TinyColor(themeConfig.value.colors.primary)
+
     const counts = Array.from(tags.value).map(([_, value]) => value.count)
     const max = Math.max(...counts)
     const min = Math.min(...counts)
@@ -31,6 +60,8 @@ export function useSakuraTags(options: {
 
   return {
     tags,
-    getTagStyle,
+    getTagStyle: getTagStyleByIndex,
+    getTagStyleByIndex,
+    getTagStyleByCount,
   }
 }