+++ theme/layouts/tags.vue
@@ -1,3 +1,13 @@
+<script setup lang="ts">
+import { defineWebPage, useSchemaOrg } from '@unhead/schema-org'
+
+useSchemaOrg([
+  defineWebPage({
+    '@type': 'CollectionPage',
+  }),
+])
+</script>
+
 <template>
-  <TagsThemeSakura />
+  <SakuraTagsLayout />
 </template>